backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
sentence-transformers
faiss-cpu
langchain
langchain-community
openai
tqdm
python-dotenv
groq
